Adiciona a funcionalidade ResponseEntity para melhor resposta dos metodos https

+++ src/main/java/dev/java10x/cadastrodeninjas/Missoes/MissaoController.java
@@ -1,7 +1,13 @@
 package dev.java10x.cadastrodeninjas.Missoes;
 
 import dev.java10x.cadastrodeninjas.Ninjas.NinjaDTO;
+import org.apache.coyote.Response;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @RestController
 @RequestMapping("missoes")
@@ -19,17 +25,38 @@ public class MissaoController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id){
-        missoesService.delete(id);
+    public ResponseEntity<String> delete(@PathVariable Long id){
+        if(missoesService.listId(id) != null){
+            missoesService.delete(id);
+            return ResponseEntity.ok().body("Missao deletada com sucesso !");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Id não encontrado !");
+        }
     }
 
     @PostMapping("/add")
-    public MissoesDTO add(@RequestBody MissoesDTO missoesDTO){
-        return missoesService.add(missoesDTO);
+    public ResponseEntity<String> add(@RequestBody MissoesDTO missoesDTO){
+        MissoesDTO missoes = missoesService.add(missoesDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Missao criada com sucesso!" + " " + missoes.getNome() + " " + missoes.getId());
     }
 
     @PutMapping("/alter/{id}")
-    public MissoesDTO alter(@PathVariable Long id , MissoesDTO dto){
-        return missoesService.alter(id,dto);
+    public ResponseEntity<?> alter(@PathVariable Long id , @RequestBody MissoesDTO dto) {
+        MissoesDTO missao = missoesService.alter(id, dto);
+        if (missao != null) {
+            return ResponseEntity
+                    .ok(missao);
+        }
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Missão nao encontrada !");
+        }
+
+
+    @GetMapping("list")
+    public ResponseEntity<?> list(){
+        if(missoesService.list().isEmpty()) {
+            List<MissoesDTO> missoes = missoesService.list();
+            return ResponseEntity.ok(missoes);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Não há missões cadastradas !");
     }
 }
